+++ BarkBook-API/src/main/java/org/teampraxis/BarkBook_API/BarkBookApiApplication.java
@@ -1,13 +1,13 @@
-package com.teampraxis;
+package org.teampraxis.BarkBook_API;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BookBarkApiApplication {
+public class BarkBookApiApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(BookBarkApiApplication.class, args);
+		SpringApplication.run(BarkBookApiApplication.class, args);
 	}
 
 }
