+++ BarkBook-UI/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "bookbark-ui",
+  "name": "barkbook-ui",
   "private": true,
   "version": "0.0.0",
   "type": "module",
