--- BookBark-API/src/main/resources/application.properties
@@ -1,18 +1,0 @@
-spring.application.name=BookBark-API
-
-spring.datasource.url=jdbc:mysql://localhost:3306/chooseTableName
-spring.datasource.username=choose userName
-spring.datasource.password=choose passWord
-
-# Specify the DBMS
-spring.jpa.database = MYSQL
-
-# Show or not log for each sql query
-spring.jpa.show-sql = false
-
-# Hibernate ddl auto (create, create-drop, update)
-spring.jpa.hibernate.ddl-auto = update
-
-# Use spring.jpa.properties.* for Hibernate native properties (the prefix is
-# stripped before adding them to the entity manager)
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL8Dialect
