--- BookBark-UI/src/App.css
@@ -1,6 +1,0 @@
-#root {
-  /* max-width: 1280px; */
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
