--- BookBark-UI/src/App.jsx
@@ -1,14 +1,0 @@
-import './App.css'
-
-function App() {
-
-  return (
-    <>
- 
-      <h1>Team Praxis</h1>
-
-    </>
-  )
-}
-
-export default App
